Add Prisma module and integrate @nestjs/config; update role service methods for better error handling

+++ src/role/role.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import { CreateRoleDto } from './dto/create-role.dto';
 import { UpdateRoleDto } from './dto/update-role.dto';
@@ -7,33 +7,55 @@ import { UpdateRoleDto } from './dto/update-role.dto';
 export class RoleService {
   constructor(private prisma: PrismaService) { }
 
-  create(dto: CreateRoleDto) {
+  async create(dto: CreateRoleDto) {
     return this.prisma.role.create({
       data: dto,
     });
   }
 
-  findAll() {
+  async findAll() {
     return this.prisma.role.findMany({
-      include: { users: true }, // opcional, mostra els usuaris del rol
+      include: {
+        _count: {
+          select: { users: true },
+        },
+      },
     });
   }
 
-  findOne(id: number) {
+  async findOne(id: number) {
     return this.prisma.role.findUnique({
       where: { id_role: id },
-      include: { users: true }, // opcional
     });
   }
 
-  update(id: number, dto: UpdateRoleDto) {
+  async update(id: number, dto: UpdateRoleDto) {
     return this.prisma.role.update({
       where: { id_role: id },
       data: dto,
     });
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const role = await this.prisma.role.findUnique({
+      where: { id_role: id },
+      include: {
+        _count: {
+          select: { users: true },
+        },
+      },
+    });
+
+    if (!role) {
+      throw new NotFoundException(`The role with ID ${id} does not exist.`);
+    }
+
+    if (role._count.users > 0) {
+      throw new BadRequestException(
+        `Failed to delete role "${role.code}" because it has ${role._count.users} users assigned. Please reassign the users first.`,
+      );
+    }
+
     return this.prisma.role.delete({
       where: { id_role: id },
     });
